Add tests for AppBar component seperate from AppBar index
