+++ app/components/global/AppBar/AppBar.test.js
@@ -1,82 +1,50 @@
 import React from 'react'
 import { mount } from 'enzyme'
-import { MockedProvider } from 'react-apollo/test-utils'
+import { MemoryRouter } from 'react-router-dom'
 import { ThemeProvider } from 'styled-components'
 import theme from '@elifesciences/elife-theme'
-import { MemoryRouter } from 'react-router-dom'
-// eslint-disable-next-line import/no-named-default
-import { default as AppBarWrapper } from '.'
-import { CURRENT_USER } from '../queries'
-import AppBar from './AppBar'
+import AppBar, { AppBarLink } from './AppBar'
 
-const defaultProps = {
-  defaultMenuItems: [
-    { label: 'Author guide', link: '/author-guide' },
-    { label: 'Reviewer guide', link: '/reviewer-guide' },
-    { label: 'Contact us', link: '/contact-us' },
-  ],
-  userMenuItems: [{ label: 'Dashboard', link: '/' }],
-}
-
-const mockUserData = {
-  data: {
-    currentUser: {
-      id: '111',
-      identities: [
-        {
-          name: 'Joe Bloggs',
-          email: '',
-          aff: '',
-          meta: {
-            firstName: 'Joe',
-            lastName: 'Bloggs',
-            __typename: 'ElifeIdentityMeta',
-          },
-          __typename: 'ElifeIdentity',
-        },
-      ],
-      __typename: 'User',
-    },
-  },
-}
-
-const currentUserMock = result => [
-  {
-    request: {
-      query: CURRENT_USER,
-    },
-    result,
-  },
-]
-
-const makeWrapper = (props, mocks = []) =>
+const makeWrapper = props =>
   mount(
     <ThemeProvider theme={theme}>
-      <MockedProvider addTypename mocks={mocks}>
-        <MemoryRouter>
-          <AppBarWrapper {...props} />
-        </MemoryRouter>
-      </MockedProvider>
+      <MemoryRouter>
+        <AppBar {...props} />
+      </MemoryRouter>
     </ThemeProvider>,
   )
 
 describe('AppBar', () => {
-  it('Only receives default menu links when no user object passed', async () => {
-    const wrapper = makeWrapper(defaultProps, currentUserMock({}))
-    await new Promise(resolve => {
-      setTimeout(resolve, 0)
+  it('generates the correct number of links from menuItems prop', () => {
+    let wrapper = makeWrapper({
+      user: {},
+      menuItems: [{ label: 'a', link: '/a' }, { label: 'b', link: '/b' }],
     })
-    wrapper.update()
-    expect(wrapper.find(AppBar).props().menuItems).toHaveLength(3)
+
+    expect(wrapper.find('[data-test-id="app-bar-menu"] a')).toHaveLength(2)
+    wrapper = makeWrapper({
+      user: {},
+      menuItems: [
+        { label: 'a', link: '/a' },
+        { label: 'b', link: '/b' },
+        { label: 'c', link: '/c' },
+      ],
+    })
+    expect(wrapper.find('[data-test-id="app-bar-menu"] a')).toHaveLength(3)
   })
 
-  it('Receives both default and user menu items when user object passed', async () => {
-    const wrapper = makeWrapper(defaultProps, currentUserMock(mockUserData))
+  it('correctly creates AppBarLinks from menuItems prop', () => {
+    const menuConfig = [
+      { label: 'a', link: '/a' },
+      { label: 'b', link: '/b' },
+      { label: 'c', link: '/c' },
+    ]
+    const wrapper = makeWrapper({ user: {}, menuItems: menuConfig })
 
-    await new Promise(resolve => {
-      setTimeout(resolve, 0)
+    menuConfig.forEach((config, index) => {
+      const link = wrapper.find(AppBarLink).at(index)
+      expect(link.props().to).toEqual(config.link)
+      expect(link.props().children).toEqual(config.label)
     })
-    wrapper.update()
-    expect(wrapper.find(AppBar).props().menuItems).toHaveLength(4)
   })
 })
